FiniteElementQuadratureRules

Documentation for FiniteElementQuadratureRules.

FiniteElementQuadratureRules.AffineGeometry

FiniteElementQuadratureRules.CompactQuadratureRule

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights

FiniteElementQuadratureRules.LagrangeLocalBasis

FiniteElementQuadratureRules.MultiLinearGeometry

FiniteElementQuadratureRules.QuadratureRule

FiniteElementQuadratureRules.ReferenceElement

FiniteElementQuadratureRules.barycentricCoordinates

FiniteElementQuadratureRules.getProperties

FiniteElementQuadratureRules.integrate

FiniteElementQuadratureRules.AffineGeometry — Type

AffineGeometry{T,mydim,cdom,Ω}

FiniteElementQuadratureRules.CompactQuadratureRule — Type

CompactQuadratureRule{T,Ω}

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights — Type

CompactQuadratureRuleWithWeights{T,Ω}

Compact rule that stores also weights.

FiniteElementQuadratureRules.LagrangeLocalBasis — Type

LagrangeLocalBasis{dim,Ω}

FiniteElementQuadratureRules.MultiLinearGeometry — Type

MultiLinearGeometry{T,mydim,cdim,Ω}

FiniteElementQuadratureRules.QuadratureRule — Type

QuadratureRule{T,D,Ω}

FiniteElementQuadratureRules.ReferenceElement — Type

ReferenceElement{D,Ω}

FiniteElementQuadratureRules.barycentricCoordinates — Method

barycentricCoordinates(domain::AbstractDomain, x::AbstractVector)

Transform reference element coordinates in a given domain into barycentric coordinates. This is in particular useful for simplex domains where these two coordinates differ. Also for prisms and pyramids where some coordinates parts are associated to a conical extension.

Parameters:

domain::AbstractDomain: The domain of the original coordinates

x::AbstractVector: The coordinates in the reference domain

Result: Barycentric coordinates associated to the domain.

FiniteElementQuadratureRules.getProperties — Method

getProperties(domain::AbstractDomain, points::AbstractVector, weights::AbstractVector)

Collect properties of a quadrature given as vector of points and weights. Currently, three properties are checked:

positive weights: property :positive

points strictly inside the domain: property :inside

points are inside or on the boundary of the domain: property :boundary

The function returns a vector of properties as corresponding Symbols.

FiniteElementQuadratureRules.integrate — Method

integrate(f::Function, qr::QuadratureRule)

Compute the integral ∫f(x)dx over a domain Ω=domain(qr) using a quadrature rule.