FiniteElementQuadratureRules

Documentation for FiniteElementQuadratureRules.

Base.Dict

FiniteElementQuadratureRules.AffineGeometry

FiniteElementQuadratureRules.AffineGeometry

FiniteElementQuadratureRules.BarycentricMonomials

FiniteElementQuadratureRules.BarycentricMonomials

FiniteElementQuadratureRules.CompactQuadratureRule

FiniteElementQuadratureRules.CompactQuadratureRule

FiniteElementQuadratureRules.CompactQuadratureRule

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.JacobiPolySet

FiniteElementQuadratureRules.LagrangePolySet

FiniteElementQuadratureRules.LagrangePolySet

FiniteElementQuadratureRules.LagrangePolySet

FiniteElementQuadratureRules.MonomialPolySet

FiniteElementQuadratureRules.MonomialPolySet

FiniteElementQuadratureRules.MonomialPolySet

FiniteElementQuadratureRules.MultiLinearGeometry

FiniteElementQuadratureRules.MultiLinearGeometry

FiniteElementQuadratureRules.MultiLinearGeometry

FiniteElementQuadratureRules.QuadratureRule

FiniteElementQuadratureRules.QuadratureRule

FiniteElementQuadratureRules.QuadratureRule

FiniteElementQuadratureRules.QuadratureRule

FiniteElementQuadratureRules.QuadratureRule

FiniteElementQuadratureRules.ReferenceElement

FiniteElementQuadratureRules.barycentricCoordinates

FiniteElementQuadratureRules.expand

FiniteElementQuadratureRules.expand

FiniteElementQuadratureRules.getProperties

FiniteElementQuadratureRules.getProperties

FiniteElementQuadratureRules.getWeights

FiniteElementQuadratureRules.getWeights

FiniteElementQuadratureRules.getWeights

FiniteElementQuadratureRules.getWeights

FiniteElementQuadratureRules.integrate

FiniteElementQuadratureRules.integrate

FiniteElementQuadratureRules.optimize

FiniteElementQuadratureRules.optimize

Functions exported from FiniteElementQuadratureRules:

FiniteElementQuadratureRules.barycentricCoordinates — Method

barycentricCoordinates(domain::AbstractDomain, x::AbstractVector)

Transform reference element coordinates in a given domain into barycentric coordinates. This is in particular useful for simplex domains where these two coordinates differ. Also for prisms and pyramids where some coordinates parts are associated to a conical extension.

Parameters:

domain::AbstractDomain: The domain of the original coordinates

x::AbstractVector: The coordinates in the reference domain

Result: Barycentric coordinates associated to the domain.

FiniteElementQuadratureRules.expand — Method

expand(cqr::CompactQuadratureRuleWithWeights{Ω,T})

Expand a compact rule into a full quadrature rule. This first expands the symmetry orbits to generate a list of quadrature points and combines it with the given quadrature weights.

FiniteElementQuadratureRules.expand — Method

expand(cqr::CompactQuadratureRule{Ω,T})

Expand a compact rule into a full quadrature rule. This first expands the symmetry orbits to generate a list of quadrature points and then computes the associated quadrature weights.

FiniteElementQuadratureRules.getProperties — Method

getProperties(domain::AbstractDomain, points::AbstractVector, weights::AbstractVector)

Collect properties of a quadrature given as vector of points and weights. Currently, three properties are checked:

positive weights: property :positive

points strictly inside the domain: property :inside

points are inside or on the boundary of the domain: property :boundary

The function returns a vector of properties as corresponding Symbols.

FiniteElementQuadratureRules.getProperties — Method

getProperties(qr::QuadratureRule)

Collect properties of a quadrature rule.

FiniteElementQuadratureRules.getWeights — Method

getWeights(::Type{T}, domain::AbstractDomain, degree::Integer, points::AbstractVector, orbits::AbstractVector)

Compute the quadrature weight using Float64 precision.

FiniteElementQuadratureRules.getWeights — Method

getWeights(::Type{T}, domain::AbstractDomain, degree::Integer, points::AbstractVector)

Compute the quadrature weight using Float64 precision.

FiniteElementQuadratureRules.getWeights — Method

getWeights(::Type{T}, domain::AbstractDomain, degree::Integer, points::AbstractVector, orbits::AbstractVector)

Compute the quadrature weight associated to given quadrature points for quadrature in a given domain, such that polynomial up to degree are integrated exactly. This overload takes into account that the points are associated to symmetry orbits and thus points in the same orbit share a quadrature weight. This makes the computation more stable and faster. The parameter T represents the number type to be used for the computation.

FiniteElementQuadratureRules.getWeights — Method

getWeights(::Type{T}, domain::AbstractDomain, degree::Integer, points::AbstractVector)

Compute the quadrature weight associated to given quadrature points for quadrature in a given domain, such that polynomial up to degree are integrated exactly. The parameter T represents the number type to be used for the computation.

FiniteElementQuadratureRules.integrate — Method

integrate(f::Function, qr::QuadratureRule)

Compute the integral ∫f(x)dx over a domain Ω=domain(qr) using a quadrature rule.

FiniteElementQuadratureRules.integrate — Method

integrate(m::BarycentricMonomials, Ω::AbstractSimplex)

Compute the integral of a barycentric monomial m over a generic simplex of volume 1. For a specific simplex, this integral needs to be scaled with its volume.

FiniteElementQuadratureRules.optimize — Method

optimize(qr::CompactQuadratureRuleWithWeights)

Optimize the position of the quadrature points given as symmetric orbits in the compact rule, by minimizing the quadrature residual on a set of polynomial basis functions. We use basis functions and their integrals given by a JacobiPolySet.

FiniteElementQuadratureRules.optimize — Method

optimize(qr::CompactQuadratureRule)

Optimize the position of the quadrature points given as symmetric orbits in the compact rule, by minimizing the quadrature residual on a set of polynomial basis functions. We use basis functions and their integrals given by a JacobiPolySet.

Types in module FiniteElementQuadratureRules:

Base.Dict — Type

Dict(qr::QuadratureRule, ref::String = "unknown")

Convert the given QuadratureRule into a Dict for exporting into a YAML file. The optional parameter ref refers to a bibtex key used to reference a publication where the quadrature rule is extracted from.

FiniteElementQuadratureRules.AffineGeometry — Type

AffineGeometry{Ω,mydim,cdom,T} <: AbstractGeometry

An AffineGeometry is a parametrization using an affine map of the form x → A*x + b. As an AbstractGeometry it is itself a mapping.

Example:

geo = AffineGeometry(ReferenceElement(Triangle()), [x0,x1,x2])
@assert domaintype(geo) == Triangle()
x = geo(λ)    # given λ in coordinates of the reference element

FiniteElementQuadratureRules.AffineGeometry — Method

(geo::AffineGeometry)(λ::AbstractVector)

Evaluate the affine geometry mapping geo in the reference element coordinate λ.

FiniteElementQuadratureRules.BarycentricMonomials — Type

BarycentricMonomials <: Function

A representation of a multi-dimensional monomial basis function of the form λ[1]^α[1] * λ[2]^α[2] * ... with λ the barycentric coordinate vector and α the exponents associated to the components of the coordinates.

FiniteElementQuadratureRules.BarycentricMonomials — Method

(m::BarycentricMonomials)(λ::AbstractVector)

Evaluate the monomial basis function m in a given barycentric coordinate λ, i.e., compute λ -> λ[1]^α[1] * λ[2]^α[2] * ...

FiniteElementQuadratureRules.CompactQuadratureRule — Type

CompactQuadratureRule{T,Ω}

A compact (quadrature) rule is defined in terms of a vector of symmetry orbits and their corresponding arguments, which allows to generate a list of quadrature points in the reference element and also their corresponding quadrature weights.

The rule is defined on a domain, e.g., a triangle or quad, has an expected degree, which means the maximal polynomial degree up to which the quadrature rule is exact.

FiniteElementQuadratureRules.CompactQuadratureRule — Method

CompactQuadratureRule(data::Dict)

Construct a CompactQuadratureRule from parsed YAML/Dict data using Float64.

FiniteElementQuadratureRules.CompactQuadratureRule — Method

CompactQuadratureRule(::Type{T}, data::Dict)

Construct a CompactQuadratureRuleWithWeights from a YAML/Dict of strings and string arrays. This is a convenience constructor typically used when reading a quadrature rule from a YAML file. All information is encoded in the Dict, in particular the fields

dim and region: characterizing the domain

degree: the quadrature degree

orbits: the symmetry orbits

positions or arguments: representing the arguments to the symmetry orbits

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights — Type

CompactQuadratureRuleWithWeights{T,Ω}

A compact (quadrature) rule with weights is defined in terms of a vector of symmetry orbits, their corresponding arguments, and a list of associated weights. This allows to generate a list of quadrature points in the reference element.

The rule is defined on a domain, e.g., a triangle or quad, has an expected degree, which means the maximal polynomial degree up to which the quadrature rule is exact.

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights — Method

CompactQuadratureRuleWithWeights(data::Dict)

Construct a CompactQuadratureRuleWithWeights from parsed YAML/Dict data using Float64.

FiniteElementQuadratureRules.CompactQuadratureRuleWithWeights — Method

CompactQuadratureRuleWithWeights(::Type{T}, data::Dict)

Construct a CompactQuadratureRuleWithWeights from a YAML/Dict of strings and string arrays. This is a convenience constructor typically used when reading a quadrature rule from a YAML file. All information is encoded in the Dict, in particular the fields

dim and region: characterizing the domain

degree: the quadrature degree

orbits: the symmetry orbits

positions: representing the arguments to the symmetry orbits

weights: for the quadrature weights

FiniteElementQuadratureRules.JacobiPolySet — Type

JacobiPolySet{Ω<:AbstractDomain,R<:Real}

A JacobiPolySet represents a set of polynomials on a domain Ω in terms of a set of basis polynomials, hereby given as Jacobi polynomials. As an AbstractPolySet it provides also the integral values over the domain of all basis functions.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(domain::AbstractDomain, degree::Integer)

Construct a JacobiPolySet on the given domain of given polynomial degree, with Float64 as data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Hexahedron, degree::Integer)

Construct a JacobiPolySet on the Hexahedron domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Line, degree::Integer)

Construct a JacobiPolySet on the Line domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Prism, degree::Integer)

Construct a JacobiPolySet on the Prism domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Pyramid, degree::Integer)

Construct a JacobiPolySet on the Pyramid domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Quadrilateral, degree::Integer)

Construct a JacobiPolySet on the Quadrilateral domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Tetrahedron, degree::Integer)

Construct a JacobiPolySet on the Tetrahedron domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.JacobiPolySet — Method

JacobiPolySet(::Type{T}, domain::Triangle, degree::Integer)

Construct a JacobiPolySet on the Triangle domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.LagrangePolySet — Type

LagrangePolySet{Ω<:AbstractDomain,R<:Real}

A LagrangePolySet represents a set of polynomials on a domain Ω in terms of a set of basis polynomials, hereby given as Lagrange polynomials. As an AbstractPolySet it provides also the integral values over the domain of all basis functions.

FiniteElementQuadratureRules.LagrangePolySet — Method

LagrangePolySet(domain::AbstractDomain, degree::Integer)

Construct a LagrangePolySet on the given domain of given polynomial degree, with Float64 as data type used for the integral values.

FiniteElementQuadratureRules.LagrangePolySet — Method

LagrangePolySet(::Type{T}, domain::AbstractDomain, degree::Integer)

Construct a LagrangePolySet on a given domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.MonomialPolySet — Type

MonomialPolySet{Ω<:AbstractDomain,R<:Real}

A MonomialPolySet represents a set of polynomials on a domain Ω in terms of a set of basis polynomials, hereby given as a monomial basis. As an AbstractPolySet it provides also the integral values over the domain of all basis functions.

FiniteElementQuadratureRules.MonomialPolySet — Method

MonomialPolySet(domain::AbstractDomain, degree::Integer)

Construct a MonomialPolySet on the given domain of given polynomial degree, with Float64 as data type used for the integral values.

FiniteElementQuadratureRules.MonomialPolySet — Method

MonomialPolySet(::Type{T}, domain::AbstractSimplex, degree::Integer)

Construct a MonomialPolySet on a simplex domain of given polynomial degree, with T the data type used for the integral values.

FiniteElementQuadratureRules.MultiLinearGeometry — Type

MultiLinearGeometry{T,cdim,Ω} <: AbstractGeometry

A MultiLinearGeometry is a geometry mapping defined in terms of linear Lagrange basis function associated to the domain Ω. It can be constructed from a reference element and a vector of corner coordinates.

FiniteElementQuadratureRules.MultiLinearGeometry — Method

(geo::MultiLinearGeometry)(λ::AbstractVector)

Evaluate the multilinear geometry mapping geo in the reference element coordinates λ, by linear combination of the associated Lagrange basis functions and corner vertices.

FiniteElementQuadratureRules.MultiLinearGeometry — Method

MultiLinearGeometry(ref::ReferenceElement, coordVector::AbstractVector)

Construct a MultiLinearGeometry as an affine geometry if the domain is a simplex.

FiniteElementQuadratureRules.QuadratureRule — Type

QuadratureRule{Ω,T,Point}

A QuadratureRule on a given domain Ω is a collection of points {xᵢ} in the associated reference element and weights {wᵢ}, such that for a given polynomial p the quadrature formula ∑ᵢ p(xᵢ) wᵢ = ∫p(x) dΩ is exact, up to polynomials of a certain degree.

FiniteElementQuadratureRules.QuadratureRule — Method

QuadratureRule(data::Dict)

Construct a QuadratureRule from parsed YAML/Dict data using Float64.

FiniteElementQuadratureRules.QuadratureRule — Method

QuadratureRule(domain::AbstractDomain, degree::Integer, points::Vector)

Construct a new QuadratureRule and compute the weights and properties of the rule automatically.

FiniteElementQuadratureRules.QuadratureRule — Method

QuadratureRule(domain::AbstractDomain, degree::Integer, points::Vector, weights::Vector)

Construct a new QuadratureRule and compute the properties of the rule automatically.

FiniteElementQuadratureRules.QuadratureRule — Method

QuadratureRule(::Type{T}, data::Dict)

Construct a QuadratureRule from a YAML/Dict of strings and string arrays. This is a convenience constructor typically used when reading a quadrature rule from a YAML file. All information is encoded in the Dict, in particular the fields

dim and region: characterizing the domain

degree: the quadrature degree

coordinates: representing the quadrature points

weights: for the quadrature weights

properties: characterizing properties of the quadrature rule

FiniteElementQuadratureRules.ReferenceElement — Type

ReferenceElement{Ω,Point}

A reference representation of a given domain Ω in terms of corner coordinates and a list of facets connecting the corners.